Crate compiles without tests. Missing tests and metrics.
